updated the watchlists
made an account fields component and updated the watchlist.
