+++ src/model/movie.ts
@@ -3,13 +3,13 @@ import { Genre } from "./genre";
 export interface Movie {
   id: number;
   title: string;
-  posterPath: string; // URL to the movie's poster image
-  backdropPath?: string; // URL to the movie's backdrop image (optional)
-  overview: string; // A brief summary of the movie
-  releaseDate: string; // Could also be a Date object
+  posterPath: string; 
+  backdropPath?: string; 
+  overview: string; 
+  releaseDate: string; 
   genres: Genre[];
-  voteAverage: number; // Average rating from 0 to 10
-  voteCount: number; // Number of votes the movie has received
+  voteAverage: number; 
+  voteCount: number; 
   originalLanguage: string;
-  // Additional properties as needed
+
 }
