--- src/pages/style.css
@@ -1,47 +1,0 @@
-body,
-html,
-#root {
-  height: 100%;
-  margin: 0;
-}
-/* .topbarContainer {
-  height: 60px;
-} */
-/* .pageContainer {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  height: 100%;
-  overflow-y: hidden;
-}
-.pageLayout {
-  display: flex;
-  flex-direction: row;
-  width: 100%;
-  height: calc(100% - 60px);
-}
-.pageContent {
-  display: flex;
-  flex-direction: column;
-  width: 85%;
-  height: 100%;
-  overflow-y: scroll;
-} */
-.searchBar {
-  width: 350px;
-  height: 20px;
-  margin: 10px;
-}
-.carouselContainer {
-  /* width: 85% !important; */
-  height: 170px;
-}
-.main-content {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  width: 100%;
-  /* height: calc(100% - 170px); */
-  margin-top: 30px;
-}
